discogs_credits: extract progress bar
Issue #32 — small atomic move: `_pInterval` / `_pPos` state + `_showBar` /
`_hideBar` to `src/progress-bar.js`. Both insertDiscogsBar (entry) and
the upcoming review-table.js extraction reach for these, so a shared
module makes sense.

~40 lines moved. Smoke (Spiritual Jazz fixture) green.
